1.JNI Memory leak fixes 2. Cache time out fix 3. Doc values memory fix 4. Other minor bugs

+++ src/main/java/org/elasticsearch/index/knn/KNNWeight.java
@@ -126,7 +126,8 @@ public class KNNWeight extends Weight {
              * neighbors we are inverting the scores.
              */
             Map<Integer, Float> scores = Arrays.stream(results).collect(
-                    Collectors.toMap(result -> result.getId(), result -> result.getScore() == 0.0f ? Float.MAX_VALUE : 1/result.getScore()));
+                    Collectors.toMap(result -> result.getId(),
+                            result -> result.getScore()==0.0f ? Integer.MAX_VALUE : 1/result.getScore()));
             int maxDoc = Collections.max(scores.keySet()) + 1;
             DocIdSetBuilder docIdSetBuilder = new DocIdSetBuilder(maxDoc);
             DocIdSetBuilder.BulkAdder setAdder = docIdSetBuilder.grow(maxDoc);
